+++ tests/Functional/JMSFunctionalTest.php
@@ -456,6 +456,19 @@ class JMSFunctionalTest extends WebTestCase
         ], json_decode($this->getModel('JMSAbstractUser')->toJson(), true));
     }
 
+    public function testIgnoredProperty(): void
+    {
+        self::assertEquals([
+            'schema' => 'JMSIgnoredProperty',
+            'type' => 'object',
+            'properties' => [
+                'regular_property' => [
+                    'type' => 'string',
+                ],
+            ],
+        ], json_decode($this->getModel('JMSIgnoredProperty')->toJson(), true));
+    }
+
     /**
      * @param array<mixed> $options
      */